Move TestDraftRegistrationApprovals to test_sanctions
